refactor: move agent-compose to lib/infra/
Move 3 source files + 1 test file from lib/agent-compose/ to
lib/infra/agent-compose/ and update 14 import paths (4 internal + 10 external).

Part of Epic #7697 — restructure src/lib/ into layered architecture.

Closes #7721
