+++ turbo/apps/web/app/api/agent/composes/[id]/route.ts
@@ -15,7 +15,7 @@ import {
   getComposeById,
   getComposeOrgId,
   deleteCompose,
-} from "../../../../../src/lib/agent-compose/compose-service";
+} from "../../../../../src/lib/infra/agent-compose/compose-service";
 import { isNotFound, isConflict } from "../../../../../src/lib/errors";
 
 const router = tsr.router(composesByIdContract, {
