+++ turbo/apps/web/app/api/telegram/register/route.ts
@@ -13,7 +13,7 @@ import {
 } from "../../../../src/lib/telegram/client";
 import { encryptSecretValue } from "../../../../src/lib/shared/crypto/secrets-encryption";
 import { generateCallbackSecret } from "../../../../src/lib/callback/hmac";
-import { resolveDefaultAgentComposeId } from "../../../../src/lib/agent-compose/resolve-default";
+import { resolveDefaultAgentComposeId } from "../../../../src/lib/infra/agent-compose/resolve-default";
 import { logger } from "../../../../src/lib/logger";
 import { checkTelegramDomain } from "../../../../src/lib/telegram/check-domain";
