+++ turbo/apps/web/src/lib/infra/agent-compose/__tests__/content-hash.test.ts
@@ -9,7 +9,7 @@ import {
   DEFAULT_VERSION_DISPLAY_LENGTH,
   MIN_VERSION_PREFIX_LENGTH,
 } from "../content-hash";
-import type { AgentComposeYaml } from "../../../types/agent-compose";
+import type { AgentComposeYaml } from "../../../../types/agent-compose";
 
 describe("content-hash", () => {
   describe("computeComposeVersionId", () => {
