+++ turbo/apps/web/src/lib/infra/agent-compose/compose-service.ts
@@ -2,14 +2,14 @@ import { eq, and, inArray } from "drizzle-orm";
 import {
   agentComposes,
   agentComposeVersions,
-} from "../../db/schema/agent-compose";
-import { storages } from "../../db/schema/storage";
-import { agentRuns } from "../../db/schema/agent-run";
+} from "../../../db/schema/agent-compose";
+import { storages } from "../../../db/schema/storage";
+import { agentRuns } from "../../../db/schema/agent-run";
 import { getInstructionsStorageName } from "@vm0/core";
-import { notFound, conflict } from "../errors";
-import { canAccessCompose } from "../agent/compose-access";
-import { listS3Objects, deleteS3Objects } from "../s3/s3-client";
-import type { AgentComposeYaml } from "../../types/agent-compose";
+import { notFound, conflict } from "../../errors";
+import { canAccessCompose } from "../../agent/compose-access";
+import { listS3Objects, deleteS3Objects } from "../../s3/s3-client";
+import type { AgentComposeYaml } from "../../../types/agent-compose";
 import type { ComposeResponse } from "@vm0/core";
 
 /**
